rizzler_server/requirements.txt
aiodns==3.4.0
aiohttp==3.9.5
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
attrs==25.3.0
beautifulsoup4==4.12.3
certifi==2025.4.26
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cryptography==42.0.8
dataclasses-json==0.6.7
dnspython==2.6.1
duckduckgo-search==6.0.0
email-validator==2.2.0
executing==2.2.0
faiss-cpu==1.8.0.post1
fastapi==0.111.1
fastapi-cli==0.0.7
filelock==3.18.0
frozenlist==1.6.0
fsspec==2025.5.1
greenlet==3.2.2
h11==0.16.0
hf-xet==1.1.3
html5lib==1.1
httpcore==1.0.9
httptools==0.6.4
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.32.4
icecream==2.1.4
idna==3.10
itsdangerous==2.2.0
jinja2==3.1.6
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.25
langchain-community==0.3.24
langchain-core==0.3.60
langchain-huggingface==0.2.0
langchain-ollama==0.2.2rc1
langchain-text-splitters==0.3.8
langsmith==0.3.42
lxml==5.4.0
markdown-it-py==3.0.0
markdownify==1.1.0
markupsafe==3.0.2
marshmallow==3.26.1
mcp==1.9.1
mdurl==0.1.2
mpmath==1.3.0
multidict==6.4.4
mypy-extensions==1.1.0
networkx==3.5
numpy==1.26.4
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
ollama==0.3.3
orjson==3.10.18
outcome==1.3.0.post0
packaging==24.2
pillow==11.2.1
propcache==0.3.1
protego==0.4.0
pycares==4.8.0
pycparser==2.22
pydantic==2.7.4
pydantic-core==2.18.4
pydantic-settings==2.9.1
pygments==2.19.1
pyreqwest-impersonate==0.5.3
pysocks==1.7.1
python-dotenv==1.0.1
python-multipart==0.0.20
python-nmap==0.7.1
pyyaml==6.0.2
readabilipy==0.3.0
regex==2024.11.6
reportlab==4.4.1
requests==2.31.0
requests-toolbelt==1.0.0
rich==14.0.0
rich-toolkit==0.14.6
safetensors==0.5.3
scapy==2.5.0
scikit-learn==1.6.1
scipy==1.15.3
selenium==4.20.0
sentence-transformers==4.1.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
sqlalchemy==2.0.41
sse-starlette==2.0.0
starlette==0.37.2
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.7.0
tqdm==4.66.6
transformers==4.52.4
trio==0.30.0
trio-websocket==0.12.2
triton==3.3.0
typer==0.15.3
typing-extensions==4.13.2
typing-inspect==0.9.0
typing-inspection==0.4.1
urllib3==2.4.0
uvicorn==0.29.0
uvloop==0.21.0
watchfiles==1.0.5
webdriver-manager==4.0.2
webencodings==0.5.1
websockets==15.0.1
wsproto==1.2.0
yarl==1.20.0
zstandard==0.23.0
